Drop `yum_requirements.txt`
This is no longer needed as we include a Fortran compiler in the
`condaforge/linux-anvil` image. Also with the new compilers a Fortran
compiler will be installed thanks to the `build` dependency. So go ahead
and drop the `yum_requirements.txt` file now that `yum` is not needed to
install a Fortran compiler.